app/src/main/java/com/azalia/submission1/ui/detail/movie/DetailMovieViewModel.kt
package com.azalia.submission1.ui.detail.movie

import androidx.lifecycle.ViewModel
import com.azalia.submission1.data.movie.MovieEntity
import com.azalia.submission1.utils.DataDummy

class DetailMovieViewModel : ViewModel() {
    private lateinit var movieId: String

    fun setSelectedMovie(movieId: String) {
        this.movieId = movieId
    }

    fun getMovie(): MovieEntity {
        lateinit var movie: MovieEntity
        val movieEntities = DataDummy.getMovies()
        for (movieEntity in movieEntities) {
            if (movieEntity.movieId == movieId) {
                movie = movieEntity
            }
        }
        return movie
    }
}